put shared fixtures to conftest.py
